backend/requirements.txt
# =============================================================================
# requirements.txt
# Bu dosya projenin ihtiyaç duyduğu Python kütüphanelerini listeler.
# Kurulum: pip install -r requirements.txt
# =============================================================================

# Web Framework
fastapi==0.115.6          # Modern, hızlı Python web framework

# ASGI Sunucusu (FastAPI'yi çalıştırmak için)
uvicorn[standard]==0.32.1 # Production-ready ASGI sunucusu

# Veritabanı
sqlalchemy==2.0.36         # Python ORM - veritabanı ile nesne tabanlı çalışma
psycopg2-binary==2.9.10    # PostgreSQL için Python driver

# Veri Doğrulama
pydantic==2.10.3           # FastAPI'nin veri doğrulama motoru
email-validator==2.2.0     # EmailStr tipi için gerekli (pydantic[email] bağımlılığı)

# Ortam Değişkenleri
python-dotenv==1.0.1       # .env dosyasını okumak için

# Dosya Yükleme (Sprint 1 - upload_image endpoint için)
python-multipart==0.0.20   # UploadFile desteği için FastAPI'nin gerektirdiği kütüphane
